+++ frontend/src/components/portfolios/portfolio/portfolio.tsx
@@ -1,11 +1,14 @@
 import React, { useState } from 'react';
 import { Portfolio as PortfolioModel, Holding } from '../../../types/portfolio';
-import { StockTransaction } from '../../../types/transaction';
-import { Card, Col, Row, Typography } from '../../../shared';
+import { StockTransaction, OptionTransaction, Transaction } from '../../../types/transaction';
+import { Alert, Card, Col, Form, FormControlProps, Row, Typography } from '../../../shared';
 import { PortfolioAPI } from '../../../api/portfolio-api';
-import TransactionsTable from '../../transactions/transactions-table/transactions-table';
+import StockTransactionsTable from '../../transactions/stock-transactions-table/stock-transactions-table';
+import OptionTransactionsTable from '../../transactions/option-transactions-table/option-transactions-table';
 import StockTransactionForm from '../../transactions/stock-transaction-form/stock-transaction-form';
-import { Alert } from 'react-bootstrap';
+import OptionTransactionForm from '../../transactions/option-transaction-form/option-transaction-form';
+import { StockOrOption } from '../../../../../src/lib/types';
+
 
 export interface PortfolioProps {
   portfolio: PortfolioModel;
@@ -17,9 +20,17 @@ export function Portfolio(props: PortfolioProps) {
   const { portfolio } = props;
   const holdings = portfolio ? Object.values(portfolio.holdings) : [];
 
+  const [transactionFormType, setTransactionFormType] = useState<StockOrOption>('stock');
   const [stockTransactionFormValue, setStockTransactionFormValue] = useState(new StockTransaction());
+  const [optionTransactionFormValue, setOptionTransactionFormValue] = useState(new OptionTransaction());
 
-  const handleTransactionFormSubmit = (value: StockTransaction) => {
+  // const handleTransactionFormTypeChange = (ev: React.FormEvent) => {
+  const handleTransactionFormTypeChange = (ev: React.FormEvent<FormControlProps>) => {
+    const value = ev.currentTarget.value as StockOrOption;
+    setTransactionFormType(value);
+  }
+
+  const handleTransactionFormSubmit = (value: Transaction) => {
     console.log('submitted:', value)
 
     PortfolioAPI.addTradeToPortfolio(value, portfolio._id, (p: PortfolioModel) => {
@@ -51,17 +62,32 @@ export function Portfolio(props: PortfolioProps) {
               {holdings.length ? (<>
                 {holdings.map((holding: Holding, i: number) => (
                   <Col key={holding.symbol}>
-                    <Card>
-                      <Card.Header>
-                        <Typography>{holding.symbol}</Typography>
-                      </Card.Header>
-                      <Card.Body>
-                        <TransactionsTable
-                          transactions={holding.transactions}
-                          onDeleteTransaction={handleDeleteTransaction}
-                        />
-                      </Card.Body>
-                    </Card>
+                    {holding.stockTransactions.length ? (
+                      <Card>
+                        <Card.Header>
+                          <Typography>{holding.symbol} (Stock)</Typography>
+                        </Card.Header>
+                        <Card.Body>
+                          <StockTransactionsTable
+                            transactions={holding.stockTransactions}
+                            onDeleteTransaction={handleDeleteTransaction}
+                          />
+                        </Card.Body>
+                      </Card>
+                    ) : null}
+                    {(holding.optionTransactions.length) ? (
+                      <Card>
+                        <Card.Header>
+                          <Typography>{holding.symbol} (Option)</Typography>
+                        </Card.Header>
+                        <Card.Body>
+                          <OptionTransactionsTable
+                            transactions={holding.optionTransactions}
+                            onDeleteTransaction={handleDeleteTransaction}
+                          />
+                        </Card.Body>
+                      </Card>
+                    ) : null}
                   </Col>
                 ))}
               </>) : (
@@ -85,16 +111,31 @@ export function Portfolio(props: PortfolioProps) {
             <Row>
               <Col>
                 <Typography variant='h6'>
-                  New Transaction
+                  <Form.Control
+                    as={'select'}
+                    onChange={handleTransactionFormTypeChange}
+                    value={transactionFormType}
+                  >
+                    <option value='stock'>New Stock Transaction</option>
+                    <option value='option'>New Option Transaction</option>
+                  </Form.Control>
                 </Typography>
               </Col>
             </Row>
           </Card.Header>
           <Card.Body>
-            <StockTransactionForm
-              onSubmit={handleTransactionFormSubmit}
-              value={stockTransactionFormValue}
-            />
+            {(transactionFormType === 'stock') && (
+              <StockTransactionForm
+                onSubmit={handleTransactionFormSubmit}
+                value={stockTransactionFormValue}
+              />
+            )}
+            {(transactionFormType === 'option') && (
+              <OptionTransactionForm
+                onSubmit={handleTransactionFormSubmit}
+                value={optionTransactionFormValue}
+              />
+            )}
           </Card.Body>
         </Card>
       </Col>